Created possibility to send messages at a certain time

+++ app/src/main/java/com/example/astrand/mappe2_s305036/entities/MyEntity.java
@@ -3,7 +3,7 @@ package com.example.astrand.mappe2_s305036.entities;
 
 
 public interface MyEntity {
-    int getId();
+    long getId();
     String info1();
     String info2();
 }
